+++ src/App.js
@@ -1,8 +1,8 @@
 import {
   Switch, Route, BrowserRouter,
 } from 'react-router-dom';
-import StockCatalogue from './components/StockCatalogue';
-import Company from './components/Company';
+import StockCatalogue from './containers/StockCatalogue';
+import Company from './containers/Company';
 
 function App() {
   return (
