+++ src/__tests__/components/Company.test.js
@@ -3,7 +3,7 @@ import { shallow, configure } from 'enzyme';
 
 import Adapter from 'enzyme-adapter-react-16';
 import store from '../../store';
-import Company from '../../components/Company';
+import Company from '../../containers/Company';
 
 configure({ adapter: new Adapter() });
 let container;
